requirements.txt
watchdog==3.0.0
pika==1.3.2
flask==2.3.3
requests==2.31.0
streamlit==1.25.0
pandas==2.0.3
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf>=3.20,<5